+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Changes
 * Texture Compression Analyzer now only shows SSIM by default to reduce clutter.
 * Detect DPS penetrators and stop them from getting merged into other meshes.
+* Merge static meshes as skinned meshes now only merges if it results in a decrease in material count.
 
 ### Bug Fixes
 * Fix ring finger to foot collider having half the size from VRChat update 2023.1.2.